minor bugs fixed in PositionIO changed eval weights to int

+++ Evaluation/EvaluateWeight.cpp
@@ -1,23 +1,23 @@
 #include "EvaluateWeight.h"
 
-const size_t EvaluateWeight::MATERIAL[2] =
+const int EvaluateWeight::MATERIAL[2] =
 {
         100, 200
 };
 
-const size_t EvaluateWeight::TEMPO[10] =
+const int EvaluateWeight::TEMPO[10] =
 {
         0, 1, 2, 3, 4, 5, 6, 7, 8, 9
 };
 
-const size_t EvaluateWeight::CENTER[5] =
+const int EvaluateWeight::CENTER[5] =
 {
         0, 1, 2, 3, 4
 };
 
-const size_t EvaluateWeight::BALANCE[5] =
+const int EvaluateWeight::BALANCE[5] =
 {
         5, 4, 3, 2, 1
 };
 
-const size_t EvaluateWeight::MOBILITY = 1;
+const int EvaluateWeight::MOBILITY = 1;
